include_file should be able to used generically
`include_file` should not assume that it will be used to include a class, nor should it assume it will be including a file in `/includes` as it could potentially be used to load files other directories (like `/views`, etc). Instead, append the `'includes/class-'` prefix in the autoloader, where it makes the most sense to assume a class in `/includes`.